AllSeen Alliance existing PDFs
These PDFs are exiting files found on the https://allseenalliance.org
web site. This repository will be used to manage the developer
portion of that web site, so these files must be copied in to
the repository in order to continue appearing on the site.

This Work Contains a Third Party Work Submitted by: Qualcomm Connected Experiences, Inc.

The third party works included in this contribution are believed to be
the property of the AllSeen Alliance, and are licensed under a
Creative Commons Attribution 4.0 International License; provided that
AllSeen Alliance source code included in these works licensed under
the ISC License per the AllSeen Alliance IP Policy.

AllJoyn is a trademark of Qualcomm Innovation Center, Inc. AllJoyn is
used here with permission to identify unmodified materials originating
in the AllJoyn open source project.

Other products and brand names may be trademarks or registered
trademarks of their respective owners.

Change-Id: If382bdab69576731e0a92d15efdee203aa29297d
Signed-off-by: Wayne Lee <waynelee@qce.qualcomm.com>