Updated the API guides for About Feature
Since the new About Feature work introduced a new set of APIs
I basically had to rewrite the API guide from scratch using
the new APIs.

We have commited to support the older APIs for at least a year
so we need to keep the older legacy documents around for user
reference. I have renamed the old documents linux-legacy and
android-legacy. I have provided links to the legacy documents
from the new documents and add a link from the TOC to the
legacy documents.

Change-Id: Ib5c7b6a0d8100dd31e4fbad69b095892367ebc1c
Signed-off-by: George Nash <georgen@qce.qualcomm.com>