Typo and formatting fixes for arduino-and-gt202
The main problems were having the first FreeRTOS download link in the
wrong section and instructing Linux users to download Eclipse for Windows.
I also adjusted some formatting for command lines and removed instructions
to custom-build python for Ubuntu 12.04. The system python for 12.04 is
already 2.7.3, and the scons package installation will install python
if it is not already there.

Change-Id: I9384de04f9f73ed8dd07ffc2e93622f5258c67ef
Signed-off-by: Mathew Martineau <mathewm@qce.qualcomm.com>