Initial commit of documentation for Security 2.0
The contents of this document, and the interfaces described,
are the result of collaborative discussions by the Security 2.0
project team.  This summary documents the final consensus of
the team.

This is a port of the AllJoyn Security 2.0 High-Level Design
document. This was ported over from revison 1.6 - draft 8
of the document.

The only work done was formating change to make sure the
Tables and images appear the same in both documents. No
content changes were made to the document.

The interface document only contains the xml specification
of the Security 2.0 interfaces. This will be expanded in a
future commit.

Change-Id: I523dbe093aeb04df544ff9cb9b4c38154e33df18
Signed-off-by: George Nash <georgen@qce.qualcomm.com>