Initial commit of documentation for Security 2.0
This is a port of the AllJoyn Security 2.0 High-Level Design
document. This was ported over from revison 1.6 - draft 8
of the document.

The only work done was formating change to make sure the
Tables and images appear the same in both documents. No
content changes were made to the document.

The interface document only contains the xml specification
of the Security 2.0 interfaces. This will be expanded in a
future commit.

Change-Id: Idc01b177bd4408fd5f674c9b14d3d97acc7f60bc
Signed-off-by: George Nash <georgen@qce.qualcomm.com>