Small organization change to Security 2.0 HLD document
Instead of placing the link to images at the bottom of the document the
link is placed at the bottom of the section the image is located in.
This will reduce the amount of scrolling the writer needs to do when
editing the document.

In addition short hand notation was used for the images. When linking to
an image if the text and the link are the same then the link name can be
omitted.  The means less text repetition which in mean less opportunity
to make typo mistakes.

Example: ![text][text] becomes ![text][]

Other minor whitespace cleanup.

Change-Id: I58f19a125239c120c3c4323dc875b816bb954ae4
Signed-off-by: George Nash <georgen@qce.qualcomm.com>