Remove RSA and PIN auth mechs from security docs
This change removes the descriptions of the RSA and PIN
authentication mechanisms, that were removed from standard client
in 15.04.  The beginning of the document is updated to say this.

The change also removes the statement that a PIN or password may be
used with ALLJOYN_ECDHE_PSK.  This is insecure and should not be
done.

Change-Id: Idea5defa0281a8f821729e0348b202ea532b989a
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>