Cleaned up documentation based on linkchecker script.
The link checker script will report broken links as well as unused
links.

All of the images that were deleted were unused.  The windows images
were empty images.

The enhanced-introspection-example.png was a image of xml.  The same
text was placed in the file using the code tag so the image could be
removed.

The rest of the documents were edited to include images that were
missed from the document. When they were initially converted to
the markdown format.

Some additional cleanup like extra spacing at the end of a lines
or tabs when spaces should have been used.
(white space changes only in documents touched for this commit.)

Change-Id: I14dabd6ad39fca250d00a56ff0bccdc91e2141ad
Signed-off-by: George Nash <georgen@qce.qualcomm.com>