ASADOC-18 API Guides are catergerized by Language not Platform
This changes the way the API Guides are categerized so they are listed
by the language the guide was for.

With the exception of Objective-C all of the langauges can be used all
any platform that supports the langauge.  C++ runs on all of the
supported platforms. The way the API guids were organized made it look
like the only platfrom that would run C++ was Linux.  The only platform
that would run Java was Android. Etc.

Change-Id: I7e28fcf40bff43748403db806f6dec46522c1e92
Signed-off-by: George Nash <georgen@qce.qualcomm.com>