Simplified ACL rule matching
Simplified access rule selection by removing priority and
enforcement explicit deny rules regardless of degree of match.
Simplifying to aid in comprehension

Corrected text describing policy generated after being claimed
to match the specified policy; outgoing messages are allowed
for ANY_TRUSTED.

Corrected generated policy to indicate there are two entries
for members under the ANY_TRUSTED ACL

Changed the description of the generated policy to indicate
it is not an example, it is the policy to be generated, and
recommend certification test be created to verify this.

Corrected a couple of places, in definition and description
of session establishment, to indicate that membership
certificate trust requies a signature chain including the
security group authority rather than any certificate authority.

Change-Id: I45e108de13d07b14a2a14034b039cff62c3f33f9
Signed-off-by: Ken Swinson <kens@qce.qualcomm.com>