ASACORE-2310 Figure 2-4 in security 2.0 HLD is incorrect
Move ConfigureWiFi call after claim process block

This is the same issue as was reviewed before (then ASADOC-51)
in https://git.allseenalliance.org/gerrit/#/c/4816
However, the wrong version (the unchanged diagram) was checked
in after fixing arrows and moving directories. That is, in that
gerrit review, patch set 3 had the right order but missed arrows.
Patch set 4 incorrectly reverted back to the original order,
and was then merged. This change includes the correct version
(with fixed arrows) as originally reviewed.

Signed-off-by: Dave Thaler <dthaler@microsoft.com>
Change-Id: I9ecdfaa42780ff7255c63554c53a32df161c0e43