CA installation and GetAllProperties
- Added permission requirements for GetAllProperties
- Removed CA as separate entry in flow diagrams.  CA is expressed
in the policy data only to avoid issue of installing CA at claim
that cannot be managed.
- Submitting remaining svg diagram sources.  Most were changed
to remove the CA entry but some are unchanged other than what occurred
as an artifact of changing file format. All png files were
regenerated from the svg even if no change was intended so there
would be a match between svg and png.  recommend-to-claim-during-onboarding.*
are only diagrams not updated in this submission.

Change-Id: I97606d19528c8241c3a6bdc4d88031cb49b7c8c6
Signed-off-by: Ken Swinson <kens@qce.qualcomm.com>