Removed the Download and Contribute pages
Removed the Download and Contribute pages for the Linux
Foundation webteam.

Since we did not know what to do with the supported platforms
page and the project life cycle page these were moved up on level.

Cleaned up the link errors.  More of the errors were links to the
download page. The should be updated once we know where the new
download page is located.

Change-Id: I4850d3a21f9949b814e957e6b49a3e9a5466d2dd
Signed-off-by: George Nash <georgen@qce.qualcomm.com>