ASACORE-2163 Removed mentions of Unity
Several locations mention the Unity plugin that is no longer supported.  The documents were scrubed removing all mentions
of Unity.  The only page that still contains a mention of Unity
is the Supported-platforms page mention the last version that the plugin was supported.

Cleaned up trailing whitespace in the edited files.

Change-Id: If139f3f9718762eaeb1d2503d42fda02cce82b1e
Signed-off-by: George Nash <georgen@qce.qualcomm.com>