Add table showing policy and manifest value needed to send msg
While developing the security 2.0 unit tests the table in the
`Enforcing the rules at message creation or receipt` section
just caused confusion when trying to figure out what values
actually needed to be set to successfully take a desired
action.

The table added to the HLD helped to clairify the end result of
all of the text listed above.  It also helps by clarifying what
is meant by sending peer/receiving peer.

The table does not change anything in the text it just helps add
clarification.

Change-Id: I540373b733323be214f371ea875b14e3ffd5a70a
Signed-off-by: George Nash <georgen@qce.qualcomm.com>